requirements.txt
Flask==2.2.5
Werkzeug==3.0.1
psycopg2-binary==2.9.9